+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Release Notes ##
 
 - 0.2.4 *Unreleased*
+    - Signal errors, warnings, etc. on macOS with emoji.
     - Update to clicore 0.5.0.
 - 0.2.3 *3 February 2026*
     - Fix version number.